refactor deepseek_ocr/fun_asr_nano generate code

+++ src/models/qwen3/config.rs
@@ -31,7 +31,7 @@ pub struct Qwen3GenerationConfig {
     pub bos_token_id: usize,
     pub pad_token_id: usize,
     pub do_sample: bool,
-    pub eos_token_id: Vec<usize>,
+    pub eos_token_id: Vec<u32>,
     pub top_p: f32,
     pub top_k: usize,
     pub temperature: f32,
